requirements.txt
pint
numpy
pyserial
